langchain-rag/hohan/requirements.txt
langchain
streamlit
PyPDF2
sentence_transformers
python-dotenv